+++ deepomatic/api/http_retry.py
@@ -21,12 +21,6 @@ class retry_if_exception_type(retry_if_exception):
         super(retry_if_exception_type, self).__init__(self.__predicate)
 
 
-class RequestsTimeout(object):
-    FAST = (3.05, 10.)
-    MEDIUM = (3.05, 60.)
-    SLOW = (3.05, 600.)
-
-
 class HTTPRetry(object):
 
     """
@@ -54,9 +48,6 @@ class HTTPRetry(object):
             :param stop (optional). Tell when to stop retrying. By default it stops retrying after a delay of 60 seconds. A last retry can be done just before this delay is reached, thus the total amount of elapsed time might be a bit higher. More details in tenacity source code https://github.com/jd/tenacity/blob/5.1.1/tenacity/stop.py
                 Raises tenacity.RetryError when timeout is reached.
             :type timeout: tenacity.stop_base
-            :param requests_timeout: timeout of each request. More details in the `requests` documentation: https://2.python-requests.org//en/master/user/advanced/#timeouts
-            :type requests_timeout: float or tuple(float, float)
-
     """
 
     class Default(object):
@@ -66,11 +57,9 @@ class HTTPRetry(object):
         RETRY_EXCEPTION_TYPES = (RequestException, )
         RETRY_EXCEPTION_TYPES_BLACKLIST = (ValueError, ProxyError, TooManyRedirects, URLRequired)
 
-    def __init__(self, retry_if=None, wait=None, stop=None,
-                 requests_timeout=RequestsTimeout.FAST):
+    def __init__(self, retry_if=None, wait=None, stop=None):
         self.retry_status_code = {}
         self.retry_if = retry_if
-        self.requests_timeout = requests_timeout
         self.wait = wait
         self.stop = stop
 
@@ -90,17 +79,7 @@ class HTTPRetry(object):
                                    wait_fixed(0.1),
                                    wait_fixed(0.1) + random_wait)
 
-    def retry(self, requests_callable, *args, **kwargs):
-        try:
-            # requests_callable must be a method from the requests module
-            # this is the timeout of requests module
-            # not the one of the retry
-            requests_timeout = kwargs.pop('timeout')
-        except KeyError:
-            requests_timeout = self.requests_timeout
-
-        functor = functools.partial(requests_callable, *args,
-                                    timeout=requests_timeout, **kwargs)
+    def retry(self, functor):
         return retry(functor, self.retry_if, self.wait, self.stop)
 
     def retry_if_status_code(self, response):
